Remove Transformer namespace
Added this namespace to make the code more extensible if other
transformation needed to be added. However, it is unlikely to have other
transomation process to be added - so removing namespace to simplify
code.
